Grabbing objects
New classes - empty GrabableObject class and Grabber class. Decision was made to not use tags like "Grabbable", because working with string format doesn't seem to be consistent and reliable. Grabber can take a GrabableObject and move it according to where the player is looking, as well as releasing it from being grabbed. Only one GrabableObject can be taken at a time.
New input action - ZoomGrabbed, which is a trigger for moving held object away or closer to the target holding position. It is activated by a mouse scroll wheel. Also, Interact input action is renamed to GrabAndRelease.
Some of the scene objects were deleted, some were rearranged on the table and got grabbability, primarly flasks. All relatively close area to the player has received mesh colliders.