Little bits of reworks
Forgot to add the floor to collidable static objects previous time, now it is fixed.
Now SubstanceParticleVisualizer only works with exposed containers, since it is only used in this case. It might be renamed or restructured later.
ObjectorSelector now has a OnObjectChanged event. Exposed containers now active when this event is triggered.
New class - SelectedContainerVisualizer, which displays current selected container substance at the bottom of the screen.