Big break = big update
Yeah...
I was working on this project bit by bit, and after 3 months of infrequent programming i have came back to it. All those bits of development combined is a lot of work done. Things are still work in progress, but i have decided to push an update anyways.
ChemicalSubstance and ChemicalReaction are now named Substance and Reaction respectively due to "Chemical" prefix use is more of a pointless.
Everything reaction related currently doesn't work since i have dedicated all of my work on substance creation and interaction, so reactions will be reworked after i am 	done with substances. For example, ReactionHandler and new ReactionVisualizer class won't work.
Main menu scene was added to preload important data before the main scene appears. The menu will have more features later. LoadLaboratoryScene and LoadMainMenu scripts were added.
MaterialSettings class was created to store specific settings for Materials that substances use. It has an internal Material field accessible via a property and can be edited through properties of settings fields. MaterialSettings must be initialized before use - either automatically after a class constructor or using a public method.
The Pair<T1, T2> class contains two fields with corresponding T1 and T2 types.
Now, instead of storing data as ScriptableObjects, data is loaded from .json files located at AppData/LocalLow/%company_name%/%app_name%. The JsonDataLoader class was created to save/load lists of data to/from .json files. This class contains internal serializable class ListHolder, which is used for easier data manipulations. For now, .json data is only loaded/saved in the new static DataStorage class, which contains list of Substances and their MaterialSettings and list of Reactions. DataStorage initialization is called on main menu to laboratory scene transition.
SubstanceVisualizer is an abstract class that is designed to handle graphical processing of chemical processes. SubstanceMeshVisualiser is responsible for visualizing substance's mesh, SelectedContainerVisualizer - displaying selected container's substance at the bottom of the screen, SubstanceParticleVisualizer - spilling animation of SubstanceExposedContainer.
InputManager was renamed to PlayerInputHolder. It is usually attached to the Player Game Object.
Application has a UI now. It is still in development, but a lot of progress has already been made. UIInputHolder is a new input class that is attached to canvas to register user input for interface interaction. At the moment there is a only one UI input - button to activate the switching of interface panel. Every seperate interface panel has new UITransitioner class attached to it. It contains UITransitioner that is activated wheni the switching button is pressed, and UnityEvent that invokes on switching. Only one of UITransitioners can be active, and only active UITransitioner will respond to UI input. UITransitionerPreload class sets first active panel on the scene.
First preloaded interface in a laboratory scene is a player interface. On switching, it deactivates player input and shows a brand new pause menu, and pause menu itself on switching reactivates player input back and switches to player interface. The other panels can mostly be accessed through usage of buttons. MessageBox is inherited from UITransitioner and can be called from anywhere. MessageBox's Game Object is located at the bottom of canvas's children.
ActivityEventTrigger is class that contains two UnityEvents that are invoked after the current GameObject is activated/deactivated.
Resetter and Resettable were added to reset UI panels after a specific event.
SubstanceEditor class and corresponding interface were created. When the panel is entered, data of this class is reset. It contains temporary Substance and MaterialSettings whose values can be edited by user. The CustomMaterialRender class is responsible for rendering MaterialSettings that is being edited. A separate camera renders 256x256 texture of sphere with current Material applied. Once a Substance is created, it is added to DataStorage and player's ControllableSubstanceContainer is filled with it.
IconStorage is another static class for storing images of substances for further use in UI. It is being initialized after DataStorage and before laboratory scene is loaded. For each Substance that is in DataStorage, it renders it in the same way as SubstanceEditor does - separate camera that renders a sphere onto a texture. Those textures are then being stored in the class.
DataStorage panel is currently under development. For now, it has experimental PanelSorter and SubstanceList classes. Icons appear as an GameObject with IconFrame class.